Move IS-related code from MetaTable to a new class InformationSchemaTable
